requirements.txt
Django>=3.2,<4.0
djangorestframework
gunicorn
psycopg2-binary
django-storages
google-cloud-storage
google-cloud-secret-manager